refactor: combine members list and invites list
